CNTRLPLANE-2589: Apply PR #945 changes for ExternalOIDC, tag renames, and test improvements
Migrate ExternalOIDC tests to OTE, rename test tags ([Operator] to [Parallel],
[OIDC] to [Serial]), update suite qualifiers with exclusion-based filtering,
add ImageStream support for known-image-checker compliance, PSA audit/warn
labels, keycloak timeout and error handling fixes, and custom
testEncryptionTypeUnset with StorageVersionMigration wait logic.
