+++ test/e2e/gitlab.go
@@ -10,7 +10,7 @@ import (
 )
 
 var _ = g.Describe("[sig-auth] authentication operator", func() {
-	g.It("[OIDC][Serial] TestGitLabAsOIDCPasswordGrantCheck", func() {
+	g.It("[Serial] TestGitLabAsOIDCPasswordGrantCheck", func() {
 		testGitLabAsOIDCPasswordGrantCheck(g.GinkgoTB())
 	})
 })
